MakeDivExpand function works correctly.
There was an initial problem with this function. The hidden button wouldn't be visible after clicking it. The problem lays in the textContent property which erases all elements if modified from a parent element. wrapping single elements into different containers solves the issue easily. The function makes possible to hide and show the full content of the todo element being clicked